requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
certifi==2020.12.5
cloudinary==1.24.0
confusable-homoglyphs==3.2.0
Django==3.1.5
django-bootstrap4==2.3.1
django-registration==3.1.1
djangorestframework==3.12.2
importlib-metadata==2.1.1
pkg-resources==0.0.0
psycopg2==2.8.6
pytz==2020.5
six==1.15.0
soupsieve==2.1
sqlparse==0.4.1
urllib3==1.26.2
zipp==3.4.0
